stCreate2 187/4 and  stRefundTest EXP Gas Calculation fix Using padded 32-byte length instead of minimal bytes for exponent  EXTCODECOPY Stack Order (OpcodeGasTable.cs)     - Bug: Stack positions were reversed (length/destOffset/offset/addr)     - Fix: Corrected to EVM spec order (addr/destOffset/offset/length)
INVALID Opcode Handling (EVMSimulator.cs)
    - Bug: INVALID only called Stop() without consuming gas or marking failure
    - Fix: Added GasRemaining = 0 and IsRevert = true

+++ src/Nethereum.EVM/Program.cs
@@ -341,7 +341,7 @@ namespace Nethereum.EVM
             var currentBytes = Memory?.Count ?? 0;
             var currentWords = (currentBytes + 31) / 32;
 
-            var requiredWords = ((int)highestAccessedByte + 31) / 32;
+            var requiredWords = (highestAccessedByte + 31) / 32;
 
             if (requiredWords <= currentWords)
                 return 0;
@@ -349,7 +349,7 @@ namespace Nethereum.EVM
             return MemoryCost(requiredWords) - MemoryCost(currentWords);
         }
 
-        private static BigInteger MemoryCost(int words)
+        private static BigInteger MemoryCost(BigInteger words)
         {
             return (words * words / 512) + (3 * words);
         }
